requirements.txt
Pillow>=9.0.0
google-cloud-storage>=2.0.0
